+++ mayan/apps/dependencies/management/commands/checkdependencies.py
@@ -1,0 +1,12 @@
+from __future__ import unicode_literals
+
+from django.core import management
+
+from ...classes import Dependency
+
+
+class Command(management.BaseCommand):
+    help = 'Output the status of the defined dependencies.'
+
+    def handle(self, *args, **options):
+        Dependency.check_all()